fix: 符号ちがいの比較を解消、あわせて配列外書きこみを修正（-Wsign-compare 3件）
int と size_t を比べると int が符号なしに変換される。3 か所とも
実際の値は非負なので結果は変わらないが、変換が起きていることを
コードに出す。

- misc1.c:128  i < sizeof(*test) * 8 → (int) を明示
- recall.c:569 known_attacks を int から uint32_t へ。0 から始めて
  増やすだけの個数で、比較相手の j が uint32_t だった
- io.c:587     while (*file != '/' && i < sizeof(user))

io.c は符号を直す過程で off-by-one も見つかった。user は
char[128]、条件が i < 128 なので i は 128 まで進みうる。ループ直後に
user[i] = '\0'; を書くので、user[128]（配列外）への書きこみが
起きていた。終端の 1 バイトを残すため上限を sizeof(user) - 1 に直した。

ふるまいの差はユーザー名が 127 文字以上のときだけで、そこは元が
スタック領域の踏み越し（未定義動作）。それ未満では同一。

+++ src/misc1.c
@@ -125,7 +125,10 @@ int randnor(int mean, int stand) {
 int bit_pos(uint32_t *test) {
     uint32_t mask = 0x1;
 
-    for (int i = 0; i < sizeof(*test) * 8; i++) {
+    // i は int、sizeof(*test) * 8 は size_t（符号なし）。そのまま比べると
+    // i が符号なしに変換される。ここは i >= 0 しか通らないので値は変わらない
+    // が、変換が起きていることを明示しておく。
+    for (int i = 0; i < (int)(sizeof(*test) * 8); i++) {
         if (*test & mask) {
             *test &= ~mask;
             return i;
